HPGL.jl

HPGL.micmeter — Method

micmeter(plotter_port; x_offset=0, y_offset=0, xtick=100, ytick=1000, xmax=10_000,
         logfile)

Continuously read from the default audio input and send commands to plot the per-buffer microphone level to plotter_port (via plot_commands!).

HPGL.plot_command! — Method

plot_command!(dest, command; pen_up_immediately_after_command)

Send single command to plot destination dest, where it will be handled by that destination type's handle_command!.

If pen_up_immediately_after_command is true, any "pen down" or "pen move while pen down" commands (PD, PA) will be followed by a "pen up command", to prevent pen bleed in situations where commands are sent infrequently to a physical pen plotter.

HPGL.plot_commands! — Method

plot_commands!(destination, commands; rate_limit_duration_sec=0.2,
               pen_up_immediately_after_command=false, logfile=missing)

Send a series of commandstodestinationvia [plotcommand!](@ref), with a pause ofratelimitdurationsecbetween each command. Iflogfileis not missing, will additionally appendcommandstologfile`.

HPGL.plot_hpgl_file! — Method

plot_hpgl_file!(destination, commands; rate_limit_duration_sec=0.2,
               pen_up_immediately_after_command=false)

Send a file of HPGL commandstodestinationvia [plotcommands!](@ref), with a pause ofratelimitdurationsec` between each command.

HPGL.start_plot_repl — Method

start_plot_repl(destination; pen_up_immediately_after_command=false,
                logfile="plotter_repl_debug_2024-05-14T21:56:14.184.hpgl")

Start REPL-like environment that prompts for individual commands and then executes them for destination via plot_command!. Additionally logs all commands entered at the REPL to logfile, unless logfile=missing.

If pen_up_immediately_after_command is false, any "pen down" or "pen move while pen down" commands (PD, PA) will be followed by a "pen up command", to prevent pen bleed in situations where commands are sent infrequently to a physical pen plotter.

Non-exported functions and types

HPGL.handle_command! — Method

handle_command!(destination::T, command)

Send command to destination, dispatching on the destination's type T. All destination types supported by this package must implement a handle_command! for that type.