HPGL.jl

HPGL.micmeter — Method

micmeter(plotter_port; x_offset=0, y_offset=0, xtick=100, ytick=1000, xmax=10_000,
         logfile)

Continuously read from the default audio input and send commands to plot the per-buffer microphone level to plotter_port (via send_plotter_cmds).

HPGL.send_plotter_cmd — Method

send_plotter_cmd(port, cmd::String; safety_up=true, logfile)

Send single cmd to plotter serial port port. If safety_up is true, any pen down/pen move instructions are followed by a "pen up command". cmd will additionally be appended to logfile path, unless logfile is missing.

HPGL.send_plotter_cmds — Method

send_plotter_cmds(port, cmds; rate_limit_duration_sec=0.2, kwargs...)

Send a series of commands via send_plotter_cmd, with a rate_limit_duration_sec pause between each function.

Non-exported functions and types